Move model and tokenizer loading into initialization for performance (#17)
Since Lambdas stay "warm" for a while and hold things in memory after a
request, this makes subsequent calls to generate more recs much faster
because they don't have to repeat the work of loading the model and
tokenizer.

This also removes model-related code from the handler, which should make
it generic across different recommenders, and consolidates it into the
recommender code, which limits the number of files experimenters will
have to edit.