Allow passing algorithm params in as query string parameters (#20)
This provides a generic way to supply algorithm parameters as part of
the request URL across a variety of algorithms. The code in `handler.py`
knows nothing about the actual recommender, but still hands off an
algorithm specific parameter (theta) that tunes the amount of
diversification. (The code in `default.py` is the algorithm-specific
part that knows which parameters to look for.)