Revise path logic and move models and data to top-level (#30)
This updates the path-location logic and rearranges the repository
structure to be more logical and reduce wasted space during Docker image
build:

- establishes top-level `src`, `data`, `models` directories for source
code, input data, and model checkpoints, respectively.
- adds `model_file_path` function to locate a model file, checking in a
`POPROX_MODELS` environment variable, the project root, and the Conda
environment prefix, in that order. This will enable POPROX model data to
be stored in a data location during deployment.
- restores `project_root` and makes it search based on the current
working directory, looking for `pyproject.toml`, so that it can work
seamlessly regardless of whether and how `poprox_recommenders` has been
installed.

It also adds tests for the `paths` module, and depends on #31 and the
docker tests to make sure none of this breaks our code in a way that can
be detected at test time.