Refactor Conda dependency specifications (#34)
This refactors the Conda dependency specifications into
`dev/environment.yml`, defining the development tools, and
`dev/constraints.yml` providing additional constraints for locking
environments.

This provides better categorization of dependencies in the lock file, so
we can install the recommenders without the dev dependencies using
Conda, and allows easier use of Conda to install only the dev deps when
just tooling but not the ability to run code is required (e.g. fetching
data to build the Docker image).

Nothing changes from the developer perspective installing their
environment from the lock file — all that changes is how we create it.