Add script to automate reliable relocking (#36)
This adds a shell script that runs `conda-lock` with the correct inputs
and environment variables to ensure that locking is successful
regardless of the user's local Conda configuration. It also tweaks a
couple of dependency specifications to make sure that the `main`
category contains a closure (`conda-lock` is not 100% on achieving
that).

Once the upstream conda-lock bug is fixed, we can update this script to
add the ability to update only our dependencies.