Break the recommendation process apart into independent-ish components (#43)
This doesn't take us all the way to a pipeline abstraction or even a
standard component interface, but it helps clarify what the various
components we already have in play do and how they fit together. There
aren't any (intentional) changes to the functionality here; this just
re-organizes the code to make it easier to think about.