Move article embedding lookup table construction to `UserEmbedder` (#52)
The lookup table is only used in the creation of user embeddings, so we
can defer constructing it until we actually need it. That avoids the
need to return two different types that contain the same information in
different formats from the `ArticleEmbedder`, which cleans up the
calling code a bit.