Standardize component argument and return types (#53)
This builds on the previous refactoring and
CCRI-POPROX/poprox-concepts#9 to standardize the component API a bit.
After this refactor, every component takes an `ArticleSet` (an
extendable wrapper around a list of articles) as the first argument and
some take an `InterestProfile` (an extendable wrapper around click
histories etc) as the second argument. Additionally, all components
return one of these two types.

The mechanism by which components extend either `ArticleSet` or
`InterestProfile` with new information that isn't part of their default
definitions is Pydantic's ability to accept and set additional
attributes beyond those in the class definition, which components can
use to create new attributes like `embeddings`, `scores`, etc. I'd like
to find a way to make these more visible, both to readers of the code
and to the code itself so that we can validate that a sequence of
components produces everything later components require before they try
to access it. For now though, I'm leaving it very implicit in the
interest of keeping the scope of this change manageable.

This doesn't take us all the way to a pipeline, but it comes pretty
close—the next step along the path is likely to execute a list of
components in sequence while forming the inputs for the next component
from a combination of the previous component's inputs and output. Once
we make it that far, then we can start to think about more complex
configurations like branching component graphs etc.