Create and apply a `RecommendationPipeline` class (#56)
This introduces a simple-ish `RecommendationPipeline` that works with
the components created in #43 and #53. Conceptually, the components run
in the order they're added to the pipeline and modify shared state that
each component can select inputs from and write outputs to. It could be
generalized (e.g. to support types other than `ArticleSet` and
`InterestProfile`), but this is enough to support the recommenders we're
currently building/using.