Make all component constructor parameters readily serializable (#83)
This moves model loading into the embedder components so that they can
take easily serializable paths as constructor arguments instead of large
binary model weights. Since we don't want to load the same model more
than once in different components, it also splits the saved model into
two files for the user and news encoders.

The model loading changes also make it easier to cache the pipelines
than the models, which then reduces the amount of dynamic pipeline
configuration we can do at request time (e.g. conditional diversifiers,
various diversifier params, the number of recommendations to request.)
As a result, we build a separate pipeline for each of the rankers and
diversifiers we want to expose in the API.


Co-authored-by: Mehtab "Shahan" Iqbal <ShahanM@users.noreply.github.com>