Add a pipeline that reranks with a softmax sampler over the article scores (#94)
This add a new `SoftmaxSampler` component and creates a recommendation
pipeline that applies it as the reranker. Since we have a framework for
gluing component together and our recommenders have a lot of overlap,
swapping it in as the reranker only takes a few lines of code beyond
what's required to implement the component itself.