Use Pixi to manage lock files (#95)
This PR uses [Pixi](https://pixi.sh/latest/) for locking and installing
locked environments instead of `conda-lock`, to (hopefully) achieve
several benefits:

- reduce custom logic required for managing lock files (pixi implements
several things we were doing manually) and implement the functionality
we were waiting on a nonexistent conda-lock bugfix for (update
`poprox-concepts` without updating the rest of the dependency stack).
- simplify installation of various environments.
- reduce error in initial environment setup by using one step to set up
a dev environment with a proper editable install.
- make environment management more consistent across POPROX repositories
(pixi is very similar to hatch in how you interact with environments and
tasks).

The primary downsides I've seen so far:

- pixi cache can get rather large.
- re-solving the repositories (including for dependency updates) can
only be done on Linux, and will create a quite large set of working
environments to complete the solve; these are hardlinked to the cache,
though, so it isn't as bad as it looks. I have also verified that the
machine on which we re-lock dependencies does not need to be
CUDA-enabled.
- because pixi hard-codes how environment paths are handled, we do have
to jump through an extra hoop to move the installed environment around,
but this isn't too bad in practice (doing it with `conda-pack`).
- everyone working on the repo will need to switch tools for installing
+ running.

With these changes, we use `pixi` for all interactions with the
environments, instead of `conda`, `conda-lock`, or `micromamba`. Pixi
seems to have more momentum than conda-lock and appears — from my
informal assessment — to be the currently-favored solution for
lock-based conda project management. So I think this change, while a
hiccup on our side, will better align our workflows and toolchains with
the direction reproducible data science / ML work appear to be going.

Unfortunately, conda-lock and pixi use different lockfile formats, and
`micromamba` can not yet install directly from a Pixi lockfile.

This change also bumps the base distribution used for the Docker image
from Amazon Linux 2 to Amazon Linux 2023 for `libc` compatibility with
Pixi's defaults.