Remove dependency on ADD --chmod working (#108)
This updates `Dockerfile` to use `curl` instead of `ADD --chmod`, which
was causing compatibility problems with Docker updates. Docker removed
support in "legacy" `docker build`, so this option only works with the
new `docker-buildx`. We'll probably be on that eventually, and many
already are, but I'd rather not have our builds break because Docker
pulled a rug, and `RUN` should keep working.

I am doing this, copying the logic from
[prefix-dev/pixi-docker](https://github.com/prefix-dev/pixi-docker/)
instead of using their images, because I don't want to deal with the
potential problems of using a Debian-based image to build and a
RedHat-based image to deploy (Amazon Linux is derived from Red Hat /
CentOS / Fedora).