Clean up dependency and project specification (#110)
This PR makes a few dependency changes and cleanups:

- Add an `eval` extra to `pyproject.toml` that includes the dependencies
of the eval scripts, so that the API documentation process doesn't error
out on importing them.
- Remove Hatch environments and scripts from `pyproject.toml` since we
are using Pixi for all of that; useful scripts I recreated as Pixi
tasks. The lint scripts I removed, since we are using pre-commit for
linting, and I didn't see value in having those definitions in two
places.
- Refactor the Pixi features and environments to add `meta` (manipulate
project files) and `eval` (run the evaluation) features, and add these
to appropriate environments (`dev` and a new `eval` environment).
- Remove `cuda` as a separate feature, and make the `eval` feature
automatically select CUDA-based PyTorch on Linux. This install works on
non-CUDA Linux machines as well (and the lockfile can be updated on
non-CUDA).

    The benefits of this change
    - We don't have as many environments to keep track of
- Installing an eval or dev environment on Linux automatically makes
acceleration available
- It's easier to manage dependency updates on multiple platforms (I can
now relock everything except `production` on a Mac)
- Updating the lock file no longer requires the space overhead of two
separate full environments (cuda and non-CUDA)
    
The biggest downside of this change is that development environment
installs on Linux always incur the space overhead of CUDA, even if the
machine does not support it. To a first approximation, this difference
looks to be about 7.5GB. Pixi uses hard links, so this space is shared
with the cache and with other Pixi environments pulling in the same
versions.