Rearrange data for access clarity and share model checkpoints + outputs (#112)
This sets us up to share data with POPROX users.

- [x] Move model checkpoints into `models/nrms-mind` subdirectory for
ease in designating sharable files
- [x] Add public and shared repositories to DVC config
- [x] Designate shared and public data
- [x] Add GitHub action to automatically update sharable data when we
update the repository

Shared data is designated with `.gitattributes`, which doesn't require
the checked files to actually be tracked in Git. I created a
`poprox-sharing` attribute, which can be either `public` (share to both
public & shared repos) or `shared` (shared repo only). The new shell
script and GitHub action automatically upload files to the appropriate
remote(s) based on their attribute marker.

Actual action debugging will need to be post-merge.

Open question: since we're playing with pipelines and actions, do we
want to include something in CI that will fail if the pipeline is not
up-to-date?