Write a file of embeddings to Parquet from eval recs generation (#124)
This grabs the candidate embeddings from each pipeline execution and
builds up a cache of embeddings over the course of the eval run. Once
the eval run finishes, it writes them to a two column Parquet file where
the first column is the article id and the second is the corresponding
embedding.