Add retargetable test fixtures and test Docker instead of Serverless Local (#153)
This makes the test suites configurable in terms of their targets to
test.

- The various `_service` fixtures will make a POPROX service available
to a test case.
- The `auto_service` fixture looks at the `POPROX_TEST_TARGET`
environment variable to see which fixture to load. Supported values are
`local` (run handlers directly), `docker-auto` (build and run a docker
container), and `docker` (connect to already-running Docker container).
Eventually it will be able to be a URL to connect to a deployed Lambda
endpoint.
- Use the tests instead of the separate test script to test the docker
container. This now tests all pipelines after building the container.
- Remove `serverless-offline` and its bits.