Update the recommender to use version 2 of the request/response format (#168)
This allows some further clean up of the main route, but more
importantly `RecommendationResponseV2` uses `RecommendationList` as the
type for the recs, which has the `extras` field that will enable
additional custom logging.

The platform side has already been updated to rely on the api version
returned by the `/warmup` route, so deploying this should transition
smoothly.