Refactor pipelines into separate configurations + cache (#169)
This is the Pipeline Config Refactor, giving us configuration "files"
and caching.

- Move each pipeline configuration into a separate file (with a `common`
file for standard elements of the configuration).
- Rewrite the pipeline load logic to scan the avialable files, instead
of using a hard-coded list, and load pipelines from there.
- Allow only the desired pipeline to be loaded, instead of loading all.
- Use a `PipelineCache` to automatically share common componens across
all loaded pipelines.
- Remove the concept of multiple pipeline names — give all pipelines one
name, the name of their module. Right now those names are unchanged from
original, I would like to rename for consistency and clarity.

With pipelines in separate files, we will be able to add (in another PR,
probaly) logic to generate and measure separately, so we only need to
re-run evaluation for the pipeline we add or change instead of
re-running _everything_.