Fix filtering by feedback type in `UserArticleFeedbackEmbedder` (#210)
This change first simplifies the code of the article feedback embedder
by inlining methods and condensing the code in the `call()` method from
there. Making the step-by-step process clearer enables some additional
logging to diagnose issues. The issue here turned out to be that the
recommender configuration was setting the feedback types to `positive`
and `negative` but the types sent in the request are actually `True` and
`False`.