Remove unnecessary direct dependencies (#216)
This removes the unnecessary direct dependencies from `pyproject.toml`,
that I think are leftovers from previous installation methods for
LensKit. In particular, it drops the direct `numpy` dep, which will let
us use NumPy 2.0, and hopefully simplify the manual site build process.

This also removes the `.python-version` file, relying on the
`requires-python` setting as the single source of truth on the required
Python version, and updates our `setup-uv` actions.