Remove MINDlarge_test data (#237)
The MIND Large Test data set does not actually have clicks / relevance
signals — it is only providing the test items for users to generate
their submissions to the MIND leaderboard.

We also weren't actually using it for anything.

This PR removes it, along with its output in training preprocessing,
from our code and workflow.