Pre-process MIND data to speed up batch inference (#240)
This refactors our MIND-based evaluation to create a DuckDB file in a
preprocessing stage. It was developed in the course of trying to get
#238 working usabley, and now subsumes and disables that PR because it
turns out expanded candidate sets won't work. However, this refactor has
a couple of benefits even without #238:

- It speeds up our evaluations and metric computations a good bit and
removes both dispatch and output bottlenecks that were hindering
evaluation scalability. Specifically, we were bottlenecked on parsing
and re-serializing test requests to dispatch work to the Ray workers,
and on serializing & deserializing data going to the output actors.
- It provides parsed versions of the MIND data with proper data types
(timestamps, lists, etc.), including splitting out the click feedback
signals, so that additional analysis and use of MIND no longer needs to
re-implement MIND parsing logic, we can just query the database tables.
This should be useful for additional analyses as we try to make sense of
what we can effectively do with this data.

Summary of changes:

- Add a `duckscript` command for automation-friendly running of DuckDB
SQL scripts, with the ability to pass variables and nice logging of
script progress.
- Create DuckDB databases for each of the MIND datasets, containing
parsed versions of both news data and “behavior” data (impression logs).
These are implemented by two SQL scripts run by `duckscript`, so that we
can have SQL syntax highlighting when we are working on those files, and
because the Python-side database logic is not specific to MIND.
- Update `MindData` to work from a read-only connection to the DuckDB
database instead of in-memory Pandas data frames. With some careful
query design, makes iterating and querying news data quite a bit faster,
and also allows for the next item.
- Separate iteration from resolving request data (`iter_profile_ids` and
`lookup_request`). With this, our parallel inference can prepare batches
that only contain the profile IDs, and defer collecting the articles for
a profile to the worker. This has two distinct benefits: first, we can
parallelize looking up article data for test profiles; second, our
previous inference code was spending an annoying amount of time
serializing and deserializing the recommendation request objects
(profiling revealed this to be a major bottleneck). When the request is
formed by the worker instead of the driver, we don't need to spend any
time serializing the request to dispatch work.
- Rework the `RecommendationWriter` design to separate _preparing_ the
data to write from actually _writing_ that data. Similarly with request
dispatch, we were sending entire recommendation and repsonse objects to
every writer, even if they only needed part of the data. Writers were
the bottleneck for inference throughput, and profiling revealed that
they were actually spending a lot of their time deserializing the
Pydantic objects instead of actually writing. By parsing out the pieces
needed for each different writer in the worker, and only serializing the
data that actually needs to be written, writers are much more efficient
and no longer seem to be a bottleneck (at least until we try to max out
a many-GPU machine with a single inference job). The object interface is
a little weird, but this is an internal object that no one else is
really touching, and I documented its methods.
- A little change to get power metering working.

The result is that mind-small inference is noticeably faster (haven't
run full yet, but expect similar improvements if not more). Somewhat
more importantly, running batch inference now keeps a fleet of Ray
workers busy on Screamer, with driver and writer processes staying well
under 100% of a CPU, which means we should also have better scalability
as we try to run on other hardware.

It also gives us a standard, processed version of MIND that we can use
for further analysis and data workflows, and to which we can add
additional data as we want.