Store tensors in shared memory for inference (#244)
This updates the batch inference logic to use shared tensor memory.

It does not currently handle multiple GPUs (or, at least, does not do
anything specific to handle them). That will be easier to handle once
Ray Direct Transport is more stable and usable. I tried to use Ray
Direct Transport, but could not get it to work with actual shared
memory. Alternatively, we could try to make manual use of Torch
Distributed, but I am hesitant to do that if we are relying on Ray as
our primary parallel compute engine.

Also cleans up a little type issue with MIND vs. POPROX evaluation data.