Add a recommender with feedback, topic filtering, and fallback (#245)
This builds on @Sris-Ty 's previous work on feedback-aware recommenders
by adding a topic filter and a fallback to the end of the pipeline. This
means that we use articles that pass a topic filter to fill out as many
slots in the final list of recs as we can and only use the
highest-scored articles among those that didn't pass as a fallback to
fill out the list if needed.