Use the V4 API when responding to recommendation requests (#246)
Based on https://github.com/CCRI-POPROX/poprox-concepts/pull/71 and the
counterpart to https://github.com/CCRI-POPROX/poprox-platform/pull/639.

This doesn't yet accomplish the eventual goal, which is to use
`Impression` as the home for all changes to how articles are displayed.
That would include changes to the headline, subhead, etc and also to
article labels (e.g. `Health`) shown somewhere in the vicinity of the
headline. It gets us closer though.