Limit topic filter to highly relevant topics
So far, we've considered all topics that AP tags an article with, even though they vary in relevance to the article.

AP assigns topics in two ways: editorially and by machine. Each method assigns a relevance score to the topic, which range from 0 to 99. Editors only assign scores of 99, 75, and 50, while the machine method assigns the full range of scores.

From Sristy's work on human labeling of article headlines, we know that some of the topic labels reflect topics that are discussed farther down in the article text and don't appear in the headline/subhead. The cutoff for that seems to be around 75, with editors substantially more likely to assign relevance scores of 75 to topics that aren't obvious from the headline.

So, this PR applies the heuristic that topics with relevance 76 or higher likely correspond to what humans would understand the topic to be from the headline/subhead. It's possible we should use this heuristic more widely, but for now the topic filter is a reasonable place to start.