Update to the latest `poprox-concepts` to fix topic prefs issue (#253)
This brings us up to date with the latest changes to our domain concept
models:
* Since we added additional entity types (e.g. `person`, `place`) to our
database and made it possible to capture preferences for them, we now
have to be more careful to use only the preferences for the entity type
we mean to. In this PR, that shows up as switching from
`.onboarding_topics` (which no longer exists) to the
`interests_by_type()` method. We also have to use `entity_interests=`
instead of `onboarding_topics=` when creating `InterestProfile` objects.
* We now have to be careful to specify `entity_type` when creating
`InterestProfile` objects, which are sometimes created manually in the
code and sometimes derived from the request data we use in our tests.
Both needed some updates.
* Imports directly from `poprox_concepts` (instead of
`poprox_concepts.domain`) are no longer available, so some imports
needed minor adjustments
* `ImpressedRecommendations`, which was introduced to shift us from
returning `Article` objects to `Impression` objects in the
recommendation response, has been renamed to `ImpressedSection`