Disambiguate the terminology used in the evaluation code (#247)
This renames a bunch of things to avoid using `profile` for anything
that isn't an actual `InterestProfile` object:
* When we're talking about the information used as inputs to
recommenders, that's now a "request". The type was already
`RecommendationRequest` so it's a straightforward replacement in this
case.
* When we're talking about what comes out of a recommender, those are
now "recommendations". For POPROX, this means newsletters. For MIND, it
means examples that similarly contain a set of impressed items.
* When we're talking about recommendations together with the
interactions on the impressed items they contain, those are now "recs
with truth". I don't love this and I'm open to other names, but it's at
least clear about what's in there.

What this reveals is that there's a gap between the two halves of our
eval data, where half of the time we're talking about request and half
of the time we're talking about recs. I'm not entirely confident I
understand how those two things are joined together at the crucial hinge
points.