Review and comment eval/generate (#254)
This is PR 1 in a series of reviewing / refactoring / cleaning up
evaluation. I started with another pass over Generate, since that's what
we had touched the most in the recent POPROX offline eval lift, and
didn't find much that still needed to be changed (before a larger
overhaul on embeddings) but added more comments, docstrings, and logging
messages to hopefully make the code easier for the next person who
touches it to follow.

This also fixes a bug where embeddings were filtered out to nothing when
running a single-process batch inference.