Update recommendation ID handling (#255)
This builds on #254 and will further clean up some bits on evaluation ID
handling and output, removing one of the remaining confusing uses of
`profile_id`:

- Call the `recommendations.parquet` output field `recommendation_id`.
- Include `recommendation_id` in the JSON-serialized recommender output.
- Standardize the eval data interface to _only_ use UUIDs, and update
`MIND` to use UUIDs for slate IDs to reduce risk of ID type errors in
client code.
- Use the slate ID as the `recommendation_id` instead of fishing an
identifier out of the interest profile. This was also causing any
metrics computed on POPROX to likely be incorrect, because it was
aggregating _all_ recommendations for a profile ID.
- Change MIND impression UUIDs to be V7 UUIDs instead of V5, so they are
sortable.