Use DuckDB for POPROX eval data (#258)
This updates the POPROX eval data loader to use DuckDB, like the MIND
loader does. This is for two reasons:

- improving performance of evaluation (removing linear scans of data
frames tables, reducing memory overhead in parallel workers)
- laying groundwork for implementing the "latest" feature in a more
straightforward manner

It also implements the `lookup_article` method to get a single article
by ID, which we had wanted.

I still need to do some further testing to make sure the code here is
correct, but if it is, my initial tests indicate a 5-10x speedup in the
offline eval generation process.

Key open question: this code and the previous code both convert the
newsletter and article timestamps to dates and compare them in order to
see if the article was available as a candidate for that newsletter.
However, they don't do anything specific with time zones, to make sure
that "same day" is the correct concept. I think we need to do one of two
things:

1. Be clear about the time zone in which we are comparing "same day",
and update the logic to do correct time zone conversion.
2. Change the candidate logic to be "was ingested in the 24 hours
preceeding the newsletter", which will be time-zone-safe.

I prefer 2, because it will be more obvious what the code is doing, and
won't rely on very subtle time zone operations to ensure and verify
correctness.

---------

Co-authored-by: Karl Higley <khigley@umn.edu>