Add --latest and --recent options to evaluation (#259)
This adds two new options to the `generate` command, along with the
plumbing in the PORPOX data loader and the DVC pipelines to make them
work:

- `--latest`: only evaluates on the last slate for each account in the
data set
- `--recent`: only evaluates on a single recent slate for each account
(specifically, the last slate at least one week before the end of the
data)

The `--recent` slates are more likely to have clicks for running
effectiveness metrics.

These are both wired into DVC (`outputs/poprox-latest` and
`outputs/poprox-recent`).

Finally, since these add the ability to run on a small set of POPROX
data, I added a POPROX measurement test to the data pipeline checks, so
our POPROX offline evals will now be tested in CI.