Rename evaluate.py to measure.py (#265)
This renames `evaluate.py` to `measure.py` so that the DVC stage names
and script name are consistent. I elected to keep the `measure` name
instead of `evaluate` because all this stage does is compute metrics
(measure the lists), not draw any conclusions or do any of the other
tasks for evaluation.